Extract callbacks into a separate class. TestCases can not be constructed outside of the framework anymore.
